+++ services/products.Domain/Itens/Commands/CreateItemCommand.cs
@@ -6,14 +6,14 @@ namespace products.Domain.Itens.Commands;
 
 public class CreateItemCommand:IRequest<NotificationResult>
 {
-    public CreateItemCommand(string? name, double price)
+    public CreateItemCommand(string? name, decimal price)
     {
         Name = name;
         Price = price;
     }
 
     public string? Name { get; set; }
-    public double Price { get; set; }
+    public decimal Price { get; set; }
 }
 public class CreateItemValidator : AbstractValidator<CreateItemCommand>
 {
