+++ services/products.Domain/Itens/Entities/Item.cs
@@ -4,7 +4,7 @@ namespace products.Domain.Itens.Entities;
 
 public class Item : Entity
 {
-    public Item(string name, double price)
+    public Item(string name, decimal price)
     {
         Name = name;
         Price = price;
@@ -13,10 +13,10 @@ public class Item : Entity
     protected Item() { }
 
     public string Name { get; private set; }
-    public double Price { get; private set; }
+    public decimal Price { get; private set; }
     public DateTime CreatedAt { get; private set; }
     public DateTime UpdatedWhen { get; private set; }
     public void UpdateName(string name) => Name = name;
-    public void UpdatePrice(double price) => Price = price;
+    public void UpdatePrice(decimal price) => Price = price;
     public void UpdatedAt() => UpdatedWhen = DateTime.UtcNow;
 }
